优化逻辑和部分注释格式

+++ src/circle_buffer.h
@@ -30,7 +30,7 @@ typedef struct
 }circle_buffer_t;
 
 
-/*
+/**
 * @brief 循环缓存初始化
 * @param cb 循环缓存指针
 * @param buffer 数据缓存地址
@@ -40,16 +40,16 @@ typedef struct
 */
 void circle_buffer_init(circle_buffer_t *cb,uint8_t *buffer,uint32_t size);
 
-/*
+/**
 * @brief 循环缓存当前数据量
 * @param cb 循环缓存指针
 * @return 循环缓存当前数据量
 * @
 * @note
 */
-uint8_t circle_buffer_size(circle_buffer_t *cb);
+uint32_t circle_buffer_size(circle_buffer_t *cb);
 
-/*
+/**
 * @brief 循环缓存刷新
 * @param cb 循环缓存指针
 * @param 
@@ -58,7 +58,7 @@ uint8_t circle_buffer_size(circle_buffer_t *cb);
 */
 uint32_t circle_buffer_flush(circle_buffer_t *cb);
 
-/*
+/**
 * @brief 循环缓存是否已满
 * @param cb 循环缓存指针
 * @return true 已满
@@ -67,7 +67,7 @@ uint32_t circle_buffer_flush(circle_buffer_t *cb);
 */
 uint8_t circle_buffer_is_full(circle_buffer_t *cb);
 
-/*
+/**
 * @brief 循环缓存是否已空
 * @param cb 循环缓存指针
 * @return true 已空
@@ -86,7 +86,7 @@ uint8_t circle_buffer_is_empty(circle_buffer_t *cb);
 */
 uint32_t circle_buffer_read(circle_buffer_t *cb,uint8_t *dst,uint32_t size);
 
-/*
+/**
 * @brief 循环缓存写入数据
 * @param cb 循环缓存指针
 * @param src 数据源地址
